update makefile for a dev usability

+++ Makefile
@@ -5,13 +5,18 @@ all: run
 .PHONY: bootstrap
 bootstrap:
 	virtualenv-2.7 --no-site-packages .
-	ln -s prod.cfg buildout.cfg
+	if ! test -f buildout.cfg;then ln -s prod.cfg buildout.cfg;fi
 	./bin/python bootstrap.py
 
 .PHONY: buildout
 buildout:
 	if ! test -f bin/buildout;then make bootstrap;fi
 	if ! test -f var/filestorage/Data.fs;then make standard-config; else bin/buildout -v;fi
+
+.PHONY: dev-install
+dev-install:
+	if ! test -f buildout.cfg;then ln -s dev.cfg buildout.cfg;fi
+	make buildout
 
 .PHONY: standard-config
 standard-config:
